fix: reset token limit for invalid symbols
