+++ apps/ai-dial-admin/src/components/EntityView/Roles/models.ts
@@ -1,0 +1,5 @@
+import { DialRoleLimits } from '@/src/models/dial/base-entity';
+
+export interface RolesGridData extends DialRoleLimits {
+  name?: string;
+}
